+++ test/test_jbpy.py
@@ -30,7 +30,11 @@ def test_roundtrip_jitc_quicklook(filename, tmp_path):
 
 EXPECTED_TRES = (
     "BLOCKA",
+    "EXOPTA",
+    "GEOPSB",
+    "ICHIPB",
     "J2KLRA",
+    "PRJPSB",
     "SECTGA",
     "STDIDC",
     "USE00A",